Handle non-switch args one at a time.
Remove args() function for handling all non-switch args as a single list.
Add arg() to handle each non-switch arg as it is encountered.  This is more
flexible and allows for handing switches in the context of other command line
arguments.

Clean up exception handling so that we no longer have to pass the arg_display string
around to all of the switch handlers.